Rename provider to terraform-provider-ad
