+++ ad/internal/ldaphelper/ldap_domain.go
