+++ ad/provider_test.go
@@ -1,4 +1,4 @@
-package msad
+package ad
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -11,6 +11,6 @@ var testAccProvider *schema.Provider
 func init() {
 	testAccProvider = Provider().(*schema.Provider)
 	testAccProviders = map[string]terraform.ResourceProvider{
-		"msad": testAccProvider,
+		"ad": testAccProvider,
 	}
 }
