+++ main.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
-	"github.com/hashicorp/terraform-provider-msad/msad"
+	"github.com/hashicorp/terraform-provider-ad/ad"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: msad.Provider})
+		ProviderFunc: ad.Provider})
 }
