+++ website/ad.erb
@@ -6,25 +6,25 @@
         <a href="/docs/providers/index.html">All Providers</a>
         </li>
 
-        <li<%= sidebar_current("docs-msad-index") %>>
-        <a href="/docs/providers/msad/index.html">MS Active Directory Provider</a>
+        <li<%= sidebar_current("docs-ad-index") %>>
+        <a href="/docs/providers/ad/index.html">MS Active Directory Provider</a>
         </li>
 
-        <li<%= sidebar_current("docs-msad-datasource") %>>
+        <li<%= sidebar_current("docs-ad-datasource") %>>
           <a href="#">Data Sources</a>
           <ul class="nav nav-visible">
-            <li<%= sidebar_current("docs-msad-datasource") %>>
-              <a href="/docs/providers/msad/d/msad_domain.html">msad_domain</a>
-              <a href="/docs/providers/msad/dmsad_user.html">msad_user</a>
+            <li<%= sidebar_current("docs-ad-datasource") %>>
+              <a href="/docs/providers/ad/d/ad_domain.html">ad_domain</a>
+              <a href="/docs/providers/ad/dad_user.html">ad_user</a>
             </li>
           </ul>
         </li>
 
-        <li<%= sidebar_current("docs-msad-resource") %>>
+        <li<%= sidebar_current("docs-ad-resource") %>>
         <a href="#">Resources</a>
         <ul class="nav nav-visible">
-          <li<%= sidebar_current("docs-msad-resource") %>>
-            <a href="/docs/providers/msad/r/msad_user.html">msad_user</a>
+          <li<%= sidebar_current("docs-ad-resource") %>>
+            <a href="/docs/providers/ad/r/ad_user.html">ad_user</a>
           </li>
         </ul>
         </li>
