+++ website/docs/r/ad_user.html.markdown
@@ -1,0 +1,26 @@
+---
+layout: "ad"
+page_title: "ad: ad_user"
+sidebar_current: "docs-ad-user"
+description: |-
+  Sample resource in the Terraform provider ad.
+---
+
+# ad_user
+
+Sample resource in the Terraform provider ad.
+
+## Example Usage
+
+```hcl
+resource "ad_user" "example" {
+  sample_attribute = "foo"
+}
+```
+
+## Argument Reference
+
+The following arguments are supported:
+
+* `sample_attribute` - Sample attribute.
+
