--- website/docs/r/msad_user.html.markdown
@@ -1,26 +1,0 @@
----
-layout: "msad"
-page_title: "msad: msad_user"
-sidebar_current: "docs-msad-user"
-description: |-
-  Sample resource in the Terraform provider msad.
----
-
-# msad_user
-
-Sample resource in the Terraform provider msad.
-
-## Example Usage
-
-```hcl
-resource "msad_user" "example" {
-  sample_attribute = "foo"
-}
-```
-
-## Argument Reference
-
-The following arguments are supported:
-
-* `sample_attribute` - Sample attribute.
-
